Blarney.Backend.NewSMT

Description

Verify circuit properties using SMT solver.

'checkAuto Info' is the one-size-fits-all verification procedure (or 'checkAuto Verbose' to get progress reports).

`checkFixed depth Info` is the same for fixed-depth (therefore more efficient).

'debugAuto Info' is not focusing on speed, but on producing minimal counter-examples.

'debugFixed depth Info' is the same for fixed-depth (therefore more efficient).

Documentation

data Verbosity

Verbosity level

Constructors

Verbose

Info

Quiet

data VerifyResult

Outcome of verification

data OutputType

Output kinds, used for filtering out unwanted parts

type Writer = OutputType -> Doc -> IO ()

Signature of functions that perform output depending on the output type

write_nothing :: Writer

Writer that never writes

write_screen :: Writer

Writer that always writes

data VerifConf

Generic Verification Config

Constructors

VerifConf

Fields

write :: Writer

giveModel :: Bool

vconfQuiet :: VerifConf

vconfDefault :: VerifConf

vconfDebug :: VerifConf

data FixedConf

Verification Config for fixed depth

Constructors

FixedConf

Fields

depth :: Int

fconfCombinational :: FixedConf

data IncrementalConf

Verification Config for incremental depth

Constructors

IncrementalConf

Fields

limit :: Maybe Int

iconfDefault :: IncrementalConf

verifyOfflineFixed :: Modular a => (Verbosity, VerifConf, FixedConf) -> a -> String -> String -> IO ()

Write SMT script to file for offline, fixed depth, verification

verifyOfflineQIFixed :: Modular a => (Verbosity, VerifConf, FixedConf) -> a -> String -> String -> IO ()

verifyLiveBounded :: Modular a => (Verbosity, VerifConf, FixedConf) -> a -> IO ()

Perform live bounded verification

verifyLiveFixed :: Modular a => (Verbosity, VerifConf, FixedConf) -> a -> IO ()

Perform live fixed depth verification

verifyLiveQIFixed :: Modular a => (Verbosity, VerifConf, FixedConf) -> a -> IO ()

verifyLiveIncremental :: Modular a => (Verbosity, VerifConf, IncrementalConf) -> a -> IO ()

Perform live incremental depth verification

checkBounded :: ProofPartRunner

Run bounded verification

checkRestrInd :: ProofPartRunner

Run restricted states induction verification

checkQuantInd :: ProofPartRunner

Run quantified induction verification

incrSeq :: Int -> Int -> [Int]

A stricty increasing sequence. n (strictly positive) controls growth rate, the bigger the slower. curr is the initial value.

proofPartGenerator :: [Int] -> Int -> ProofPartRunner -> AssertProofPartGenerator

Parametric proof part generator depths is a sequence of depths count is the number of concurrent jobs runner is the proof part runner

verifyCircuit :: Modular a => AssertProofPartGenerator -> (Verbosity, VerifConf) -> a -> IO ()

Concurrent verification

checkAuto :: Modular a => Verbosity -> a -> IO ()

Automatic verification procedure Aimed at giving a result as soon as possible

checkFixed :: Modular a => Int -> Verbosity -> a -> IO ()

Fixed depth verification procedure Aimed at giving a result as soon as possible

debugAuto :: Modular a => Verbosity -> a -> IO ()

Automatic minimal depth counterexample generation Note: Will run forever if there are no counterexamples TODO: Compare performance with incremental solution

debugFixed :: Modular a => Int -> Verbosity -> a -> IO ()

Fixed depth counterexample generation Note: Returns Unknown if there is no such counterexample