Blarney.Backend.Simulation

Description

Simulate a Blarney Netlist.

Documentation

compileSim

Arguments

:: Map String Simulator
existing Simulators (for nested Custom)

-> Netlist
Netlist for the circuit to compile a Simulator for

-> (String -> IO ())
Function to implement Blarney's display statement

-> IO Simulator

A function to turn a Netlist into a Simulator

runSim

Arguments

:: Simulator
the Simulator to simulate, compiled with compileSim

-> SignalMap
the input signals to the Simulator

-> IO SignalMap

Simulate a Simulator until a Finish Prim is triggered. This is strict in that the effects are evaluated before returning the outputs