Blarney.Backend.Verilog

Description

Convert Blarney Netlist to Verilog modules.

Documentation

genVerilogModule

Arguments

:: Netlist
Blarney netlist

-> String
Module name

-> String
Output directory

-> IO ()

Convert given Blarney Netlist to a Verilog module

genVerilogTop

Arguments

:: Netlist
Blarney module

-> String
Top-level module name

-> String
Output directory

-> IO ()

Convert given Blarney Netlist to a top-level Verilog module, and automatically generate a sample verilator wrapper and makefile too. This is useful for simple examples and projects. Most projects will probably require a more customised verilator wrapper; in that case, just use genVerilogModule, even for the top-level module.