Blarney.ClientServer

Description

This module defines the Client and Server types.

Documentation

data Client req_t resp_t

Clients produce requests and consume responses

Constructors

Client

Fields

reqs :: Source req_t

resps :: Sink resp_t

data Server req_t resp_t

Servers consume requests and produce responses

Constructors

Server

Fields

reqs :: Sink req_t

resps :: Source resp_t

debugClient :: (FShow req_t, FShow resp_t) => Client req_t resp_t -> Format -> Client req_t resp_t

Add debug displays to client's source and sink

debugServer :: (FShow req_t, FShow resp_t) => Server req_t resp_t -> Format -> Server req_t resp_t

Add debug displays to servers's source and sink

nullServer :: Bits resp_t => Server req_t resp_t

Server that ignores all requests and generates no responses

nullClient :: Bits req_t => Client req_t resp_t

Client that ignores all responses and generates no requests