Blarney.Core.BV

Description

This module represents the core of the Blarney HDL, upon which all hardware description features are built. It provides untyped bit vectors, i.e. bit vectors whose width is not specified in the type. Hardware developers should not use these untyped primitives directly (unless they know what they are doing), because width-mistmatches are not checked. Any bit-vector can be evaluated symbolically, using a feature similar to Observable Sharing [1], to produce a netlist, i.e. a graph whose nodes are primitive component instances and whose edges are connections.

K. Claessen, D. Sands. Observable Sharing for Functional Circuit Description, ASIAN 1999.

Untyped bit vectors

data BV

An untyped bit vector output wire from a primitive component instance

Constructors

BV

Fields

bvPrim :: Prim

What kind of primitive produced this bit vector?

bvInputs :: [BV]

Inputs to the primitive instance

bvOutput :: OutputName

Output

bvNameHints :: NameHints

Name (hints) characterising the BV

bvInstId :: InstId

Unique BV identifier

makePrim :: Prim -> [BV] -> [OutputName] -> [BV]

Helper function for creating an instance of a primitive component

makePrim0 :: Prim -> [BV] -> BV

Create instance of primitive component with no output

makePrim1 :: Prim -> [BV] -> BV

Create instance of primitive component which has one output

Bit-vector primitives

constBV :: Width -> Integer -> BV

Constant bit vector of given width

dontCareBV :: Width -> BV

Don't care of given width

addBV :: BV -> BV -> BV

Adder

subBV :: BV -> BV -> BV

Subtractor

mulBV :: BV -> BV -> BV

Multiplier (input and output widths are same)

fullMulBV :: Bool -> BV -> BV -> BV

Multiplier (full precision, i.e. output width is 2x input width)

divBV :: BV -> BV -> BV

Quotient

modBV :: BV -> BV -> BV

Remainder

invBV :: BV -> BV

Bitwise not

andBV :: BV -> BV -> BV

Bitwise and

orBV :: BV -> BV -> BV

Bitwise or

xorBV :: BV -> BV -> BV

Bitwise xor

leftBV :: BV -> BV -> BV

Left shift

rightBV :: BV -> BV -> BV

Right shift

arithRightBV :: BV -> BV -> BV

Right shift

equalBV :: BV -> BV -> BV

Equality comparator

notEqualBV :: BV -> BV -> BV

Disequality comparator

lessThanBV :: BV -> BV -> BV

Less-than comparator

lessThanEqBV :: BV -> BV -> BV

Less-than-or-equal comparator

replicateBV :: Int -> BV -> BV

Replicate a bit

zeroExtendBV :: OutputWidth -> BV -> BV

Zero-extend a bit vector

signExtendBV :: OutputWidth -> BV -> BV

Sign-extend a bit vector

selectBV :: (BitIndex, BitIndex) -> BV -> BV

Bit selection

concatBV :: BV -> BV -> BV

Bit vector concatenation

muxBV :: BV -> [BV] -> BV

Multiplexer

idBV :: BV -> BV

Identity function

testPlusArgsBV :: String -> BV

Test plusargs

inputPinBV :: Width -> String -> BV

Input pin (named Verilog pin)

regBV :: Width -> BV -> BV -> BV

Register of given width with initial value

regEnBV :: Width -> BV -> BV -> BV -> BV

Register of given width with initial value and enable wire

mergeWritesBV :: MergeStrategy -> Width -> [(BV, BV)] -> BV

Merge a list of inputs together

regFileReadBV :: RegFileInfo -> BV -> BV

Read from register file

getInitBV :: BV -> InitVal

Get the value of a constant bit vector, which may involve bit manipulations. Used to determine the initial value of a register.

ramBV :: OutputWidth -> Maybe String -> (BV, BV, BV, BV, Maybe BV) -> BV

Single-port block RAM. Inputs: address, data, write-enable, read-enable, optional byte-enable

dualRamBV :: OutputWidth -> Maybe String -> (BV, BV, BV, BV, BV, Maybe BV) -> BV

Simple dual-port block RAM. Inputs: read address, write address, data, write-enable, read-enable, optional byte-enable

trueDualRamBV :: OutputWidth -> Maybe String -> (BV, BV, BV, BV, Maybe BV) -> (BV, BV, BV, BV, Maybe BV) -> (BV, BV)

True dual-port block RAM. Inputs (x2): address, data, write-enable, read-enable, optional byte-enable

Other misc helpers

bvPrimOutWidth :: BV -> OutputWidth

helper to get the OutputWidth of the Prim of a BV

lookupParam :: [Param] -> String -> String

Given a parameter name, return the parameter value

addBVNameHint :: BV -> NameHint -> BV

Add a name hint to the BV

addBVNameHints :: BV -> NameHints -> BV

Add name hints to the BV