Blarney.Core.Bit

Description

This module provides size-typed bit vectors and circuit primitives, on top of Blarney's untyped bit vectors and circuit primitives. Hardware developers should always use the typed versions!

Typed bit-vectors

newtype Bit (n :: Nat)

Phantom type wrapping an untyped bit vector, capturing the bit-vector width. All bit vectors are members of the Num and Cmp classes.

Constructors

FromBV

Fields

toBV :: BV

widthOf :: KnownNat n => Bit n -> Int

Determine width of bit-vector from type

unsafeWidthOf :: Bit n -> Int

Determine width of bit-vector from underlying BV

valueOf :: forall n. KnownNat n => Int

Convert type Nat to Ingeter value

constant :: KnownNat n => Integer -> Bit n

Constant bit-vector

lit :: KnownNat n => Integer -> Bit n

nameBit :: String -> Bit n -> Bit n

Give a name to a 'Bit n' signal

testPlusArgs :: String -> Bit 1

Test plusargs

true :: Bit 1

True

false :: Bit 1

False

Bit-vector arithmetic

(.+.) :: Bit n -> Bit n -> Bit n infixl 6

Adder

add :: forall {n :: Nat}. Bit n -> Bit n -> Bit n

Adder

(.-.) :: Bit n -> Bit n -> Bit n infixl 6

Subtractor

sub :: forall {n :: Nat}. Bit n -> Bit n -> Bit n

Subtractor

(.*.) :: Bit n -> Bit n -> Bit n infixl 7

Multiplier

mul :: forall {n :: Nat}. Bit n -> Bit n -> Bit n

Multiplier

fullMul :: Bool -> Bit n -> Bit n -> Bit (2 * n)

Multiplier (full precision)

(./.) :: Bit n -> Bit n -> Bit n infixl 7

Quotient

quotient :: forall {n :: Nat}. Bit n -> Bit n -> Bit n

Quotient

(.%.) :: Bit n -> Bit n -> Bit n infixl 7

Remainder

remainder :: forall {n :: Nat}. Bit n -> Bit n -> Bit n

Remainder

Bitwise operations on bit-vectors

inv :: Bit n -> Bit n

Bitwise invert

(.&.) :: Bit n -> Bit n -> Bit n infixl 7

Bitwise and

(.|.) :: Bit n -> Bit n -> Bit n infixl 5

Bitwise or

(.^.) :: Bit n -> Bit n -> Bit n infixl 6

Bitwise xor

(.&&.) :: Bit 1 -> Bit 1 -> Bit 1 infixr 3

Logical and

(.||.) :: Bit 1 -> Bit 1 -> Bit 1 infixr 2

Logical or

(.==>.) :: Bit 1 -> Bit 1 -> Bit 1 infixr 1

Logical implication

(.<==>.) :: Bit 1 -> Bit 1 -> Bit 1 infixr 1

Logical equivalence

(.<<.) :: Bit n -> Bit m -> Bit n infixl 8

Shift left

(.>>.) :: Bit n -> Bit m -> Bit n infixl 8

Shift right

(.>>>.) :: Bit n -> Bit m -> Bit n infixl 8

Arithmetic shift right

rotl :: Bit n -> Bit m -> Bit n

Rotate left

rotr :: Bit n -> Bit m -> Bit n

Rotate right

Bit-vector comparison primitives

class Cmp a where

Minimal complete definition

Nothing

Methods

(.<.) :: a -> a -> Bit 1 infix 4

default (.<.) :: (Generic a, GCmp (Rep a)) => a -> a -> Bit 1

(.<=.) :: a -> a -> Bit 1 infix 4

default (.<=.) :: (Generic a, GCmp (Rep a)) => a -> a -> Bit 1

(.==.) :: a -> a -> Bit 1 infix 4

default (.==.) :: (Generic a, GCmp (Rep a)) => a -> a -> Bit 1

(.>.) :: a -> a -> Bit 1 infix 4

(.>=.) :: a -> a -> Bit 1 infix 4

(.!=.) :: a -> a -> Bit 1 infix 4

less :: Bit n -> Bit n -> Bit 1

Unsigned less-than

lessEq :: Bit n -> Bit n -> Bit 1

Unsigned less-than-or-equal-to

equal :: Bit n -> Bit n -> Bit 1

Equality

notEqual :: Bit n -> Bit n -> Bit 1

Disequality

signedLess :: Bit n -> Bit n -> Bit 1 infixl 8

Signed less than

signedGreater :: Bit n -> Bit n -> Bit 1 infixl 8

Signed greater than

signedLessEq :: Bit n -> Bit n -> Bit 1 infixl 8

Signed less than or equal

signedGreaterEq :: Bit n -> Bit n -> Bit 1 infixl 8

Signed greater than or equal

Bit-vector width adjustment

rep :: KnownNat n => Bit 1 -> Bit n

Replicate bit

zeroExtend :: (KnownNat m, n <= m) => Bit n -> Bit m

Zero extension

signExtend :: (KnownNat m, n <= m) => Bit n -> Bit m

Sign extension

(#) :: Bit n -> Bit m -> Bit (n + m) infixr 8

Bit-vector concatenation

upper :: (KnownNat m, m <= n) => Bit n -> Bit m

Extract most significant bits

truncateLSB :: forall m n. (KnownNat m, m <= n) => Bit n -> Bit m

Extract most significant bits

lower :: (KnownNat m, m <= n) => Bit n -> Bit m

Extract least significant bits

truncate :: forall m n. (KnownNat m, m <= n) => Bit n -> Bit m

Extract least significant bits

split :: KnownNat n => Bit (n + m) -> (Bit n, Bit m)

Split bit vector

dropBits :: forall d n. KnownNat d => Bit (d + n) -> Bit n

Drop most significant bits

dropBitsLSB :: forall d n. KnownNat n => Bit (n + d) -> Bit n

Drop least significant bits

invMSB :: Bit n -> Bit n

Invert most significant bit

Bit-vector selection primitives

slice :: forall (hi :: Nat) (lo :: Nat) i o. (KnownNat hi, KnownNat lo, (lo + o) ~ (hi + 1), (hi + 1) <= i, o <= i) => Bit i -> Bit o

Statically-typed bit selection. Use type application to specify upper and lower indices.

untypedSlice :: KnownNat m => (Int, Int) -> Bit n -> Bit m

Dynamically-typed bit selection

unsafeSlice :: (Int, Int) -> Bit n -> Bit m

Untyped bit selection (try to avoid!)

at :: forall (i :: Nat) n. (KnownNat i, (i + 1) <= n) => Bit n -> Bit 1

Statically-typed bit indexing. Use type application to specify index.

untypedAt :: Int -> Bit n -> Bit 1

Dynamically-typed bit indexing

unsafeAt :: Int -> Bit n -> Bit 1

Dynamically-typed bit indexing

Bit-vector inputs

inputPin :: forall n. KnownNat n => String -> Bit n

Input bit-vector

Bit-vector registers

reg :: Bit n -> Bit n -> Bit n

Register

regEn :: Bit n -> Bit 1 -> Bit n -> Bit n

Register with enable wire

mergeWritesBit :: MergeStrategy -> Width -> [(Bit 1, Bit n)] -> Bit n

Merge inputs together according to a merging strategy

Misc. bit-vector operations

mux :: Bit 1 -> Bit n -> Bit n -> Bit n

Two-input multiplexer

nmux :: Bit w -> [Bit n] -> Bit n

Multiplexer using a selector signal to index a list of input signals. Raises a circuit generation time error on empty list of inputs

liftNat :: Int -> (forall n. KnownNat n => Proxy n -> a) -> a

Lift integer value to type-level natural

fromBitList :: KnownNat n => [Bit 1] -> Bit n

Convert list of bits to bit vector

toBitList :: KnownNat n => Bit n -> [Bit 1]

Convert bit vector to list of bits

newtype OneHotList

One-hot bit list

Constructors

OneHotList [Bit 1]

unsafeFromBitList :: [Bit 1] -> Bit n

Collapse a '[Bit 1]' of size n to a single 'Bit n'

unsafeToBitList :: Bit n -> [Bit 1]

Expand a single 'Bit n' to a '[Bit 1]' of size n

onBitList :: ([Bit 1] -> [Bit 1]) -> Bit n -> Bit n

Apply bit-list transformation on bit-vector

unsafeBitCast :: Bit n -> Bit m

Cast one size of vector to another (dangerous)

truncateCast :: (KnownNat n, KnownNat m) => Bit n -> Bit m

Similar to truncate but width check done at elaboration time

truncateLSBCast :: (KnownNat n, KnownNat m) => Bit n -> Bit m

Similar to truncateLSB but width check done at elaboration time

zeroExtendCast :: (KnownNat n, KnownNat m) => Bit n -> Bit m

Similar to zeroExtend but width check done at elaboration time

cast :: (KnownNat n, KnownNat m) => Bit n -> Bit m

Zero extend or truncate input to give output

class GCmp f where

For generic deriving of Cmp class

Methods

gCmpLT :: f a -> f a -> Bit 1

gCmpLTE :: f a -> f a -> Bit 1

gCmpEQ :: f a -> f a -> Bit 1

Signed typed bit-vectors

newtype Signed (n :: Nat)

Signed bit vectors

Constructors

Signed (Bit n)

toSigned :: Bit n -> Signed n

Convert to signed bit-vector

fromSigned :: Signed n -> Bit n

Convert from signed bit-vector