Blarney.Core.Bits

Description

Any type in the Bits class can be represented in hardware at runtime. For example, values of a type in the Bits class can be stored in a register, or carried along a wire. The Bits class supports generic deriving.

Documentation

class Bits a where

Minimal complete definition

Nothing

Associated Types

type SizeOf a :: Nat

Type-level size of bit-vector

type SizeOf a = GSizeOf (Rep a)

Methods

sizeOf :: a -> Int

Value-level size of bit-vector

default sizeOf :: (Generic a, GBits (Rep a), GSizeOf (Rep a) ~ SizeOf a) => a -> Int

pack :: a -> Bit (SizeOf a)

Convert to a bit-vector

default pack :: (Generic a, GBits (Rep a), GSizeOf (Rep a) ~ SizeOf a) => a -> Bit (SizeOf a)

unpack :: Bit (SizeOf a) -> a

Convert from a bit-vector

default unpack :: (Generic a, GBits (Rep a), GSizeOf (Rep a) ~ SizeOf a) => Bit (SizeOf a) -> a

nameBits :: String -> a -> a

Name a Bits

default nameBits :: (Generic a, GBits (Rep a), GSizeOf (Rep a) ~ SizeOf a) => String -> a -> a

class GBits f where

Associated Types

type GSizeOf f :: Nat

Methods

gsizeOf :: f a -> Int

gpack :: f a -> Bit (GSizeOf f)

gunpack :: Bit (GSizeOf f) -> f a

gnameBits :: String -> f a -> f a