Blarney.Core.FShow

Description

Blarney's display statement can display values of any types in the FShow class. The FShow class supports generic deriving.

Documentation

data Format

Format for displaying values in simulation

Constructors

Format [FormatItem]

type FormatItem = (DisplayArg, Maybe BV)

A format item is a display arg, tagged with a bit vector

formatString :: String -> Format

Convert a string to a format

formatBit :: DisplayArgRadix -> Maybe Int -> Bit n -> Format

Convert bit vector to a format

formatCond :: Bit 1 -> Format -> Format

Conditional format (empty if condition fails)

formatBin :: Int -> Bit n -> Format

Format bit vector in binary with given amount of zero padding

formatDec :: Int -> Bit n -> Format

Format bit vector in decimal with given amount of zero padding

formatHex :: Int -> Bit n -> Format

Format bit vector in hex with given amount of zero padding

class FShow a where

Minimal complete definition

Nothing

Methods

fshow :: a -> Format

default fshow :: (Generic a, GFShow (Rep a)) => a -> Format

fshowList :: [a] -> Format

class GFShow f where

Methods

gfshow :: Bool -> f a -> Format