Blarney.Core.Interface

Description

Rather than flattening the entire design hierarchy down to a single netlist, we want to maintain the modular structure when generating external modules. Interface allows Blarney functions to be turned into external modules, and external modules to be instantiated in a Blarney description.

Documentation

class Interface a where

Minimal complete definition

Nothing

Methods

toIfc :: a -> (IfcTerm, IfcType)

For generic deriving

default toIfc :: (Generic a, GInterface (Rep a)) => a -> (IfcTerm, IfcType)

fromIfc :: IfcTerm -> a

default fromIfc :: (Generic a, GInterface (Rep a)) => IfcTerm -> a

data IfcTerm

Constructors

IfcTermUnit

IfcTermBV BV

IfcTermAction (Action IfcTerm)

IfcTermProduct IfcTerm IfcTerm

IfcTermFun (IfcTerm -> IfcTerm)

data IfcType

Constructors

IfcTypeUnit

IfcTypeBV Width

IfcTypeAction IfcType

IfcTypeProduct IfcType IfcType

IfcTypeFun IfcType IfcType

IfcTypeField PortInfo IfcType

toIfcTerm :: Interface a => a -> IfcTerm

fromIfcTerm :: Interface a => IfcTerm -> a

toIfcType :: Interface a => a -> IfcType

toPorts :: ToPorts a => a

fromPorts :: FromPorts a b => IfcTerm -> a -> b

data PortInfo

The user can request certain properties, e.g. names, in a generated flat interface (e.g. Verilog). Such properties are captured per-port as:

Constructors

PortInfo

Fields

name :: String

Desired field name

argNames :: [String]

Desired method argument names

alwaysEnabled :: Bool

Avoid generation of method enable wire

skipName :: Bool

Do not generate a name for this port. Useful when we are only intested in a method's arguments. Use with care; this can introduce ambiguous names in generated interface.

enableName :: String

Name extension for a method's enable wire

portEmpty :: PortInfo

Empty information about a port

portName :: String -> PortInfo

Name a port

portMethod :: String -> [String] -> PortInfo

Name a method port, i.e. the method and its args

portMethodEn :: String -> String -> [String] -> PortInfo

Name a method port and its enable wire

portMethodAlwaysEn :: String -> [String] -> PortInfo

Name an always-enabled method port

portSkipName :: PortInfo

Skip name augmentation for a port

class Method a where

This class defines which functions make a valid Interface. Specifially, a Method is any function returning an Action whose arguments are in Interface and Bits. Note that we don't actually use the Bits dictionary, but it captures at the type level the restriction that method arguments must be assignable at the circuit level.

Methods

toMethodTerm :: a -> IfcTerm

fromMethodTerm :: IfcTerm -> a

toMethodType :: a -> IfcType

class Modular a where

Methods

makeMod :: Int -> a -> Declare ()

makeInst :: String -> InstanceInfo -> Int -> Declare () -> Maybe CustomNetlist -> a

data InstanceInfo

Constructors

InstanceInfo

Fields

instanceParams :: [Param]

instanceClock :: Maybe Clock

instanceReset :: Maybe Reset

makeModule :: Modular a => a -> Module ()

makeInstanceWithInfo

Arguments

:: Modular a

=> String
Name of component being instantiated

-> InstanceInfo
Instance information

-> Maybe CustomNetlist
Optional netlist for component

-> a

makeInstance :: Modular a => String -> a

makeBoundaryWithInfo :: Modular a => InstanceInfo -> String -> a -> a

Make a named synthesis boundary around the given module. Note that it is possible for the supplied module to capture data that is independent of the module's inputs; this may or may not be desirable, so take care.

makeBoundary :: Modular a => String -> a -> a

makeBoundaryWithClockAndReset :: Modular a => (Clock, Reset) -> String -> a -> a