Blarney.Core.Net

Description

This module provides types and functions to represent circuits as Netlists that can then be rendered as Verilog or in other formats...

Documentation

data Net

Net type representing a Netlist node

Constructors

Net

Fields

netPrim :: Prim

The Net 's Primitive

netInstId :: InstId

The Net 's InstId identifier

netInputs :: [NetInput]

The Net 's list of NetInput inputs

netNameHints :: NameHints

The Net 's NameHints

type WireId = (InstId, OutputName)

A WireId uniquely identify a wire with a Net's instance identifier (InstId) and an output name (OutputName)

data NetInput

A Net's input (NetInput) can be: - a wire, using the InputWire constructor - a complex expression, using the InputTree constructor

Constructors

InputWire WireId

InputTree Prim [NetInput]

netInputWireIds :: NetInput -> [WireId]

Helper function to extract all NetInput's WireIds

remapNetInputInstId :: (InstId -> InstId) -> NetInput -> NetInput

Helper function remap the InstIds of a NetInput

remapNetInstId :: (InstId -> InstId) -> Net -> Net

Helper function remap the InstIds of a Net

type Netlist = Array InstId Net

A Netlist, represented as an 'Array InstId Net'

class Monad m => ToNetlist a m where

Methods

toNetlist :: a -> m Netlist

getNet :: Netlist -> InstId -> Net

Extract the Net from a Netlist at the provided InstId. Raise an error if no Net with this InstId is present.

topologicalSort :: Netlist -> [InstId]

get a topologically sorted '[InstId]' for the given Netlist

partialTopologicalSort :: Netlist -> InstId -> [InstId]

get a partially topologically sorted '[InstId]' for the subset of the given Netlist from the given Net's InstId