Blarney.Core.NetHelpers

Description

This module provides various functions over Nets and Netlists. See Prim for the definition of Net and Netlist. (The definitions were moved there to break an import cycle.)

Documentation

netInputWireIds :: NetInput -> [WireId]

Helper function to extract all NetInput's WireIds

remapNetInputInstId :: (InstId -> InstId) -> NetInput -> NetInput

Helper function remap the InstIds of a NetInput

remapNetInstId :: (InstId -> InstId) -> Net -> Net

Helper function remap the InstIds of a Net

class ToNetlist a where

Methods

toNetlist :: a -> Netlist

getNet :: Netlist -> InstId -> Net

Extract the Net from a Netlist at the provided InstId. Raise an error if no Net with this InstId is present.

topologicalSort :: Netlist -> [InstId]

get a topologically sorted '[InstId]' for the given Netlist

partialTopologicalSort :: Netlist -> InstId -> [InstId]

get a partially topologically sorted '[InstId]' for the subset of the given Netlist from the given Net's InstId