Blarney.Core.RAM

Description

Block RAM primitives

ram :: (Bits a, Bits d) => (a, d, Bit 1, Bit 1) -> d

Uninitialised block RAM. (Read during write: reads "dont care")

ramInit :: (Bits a, Bits d) => String -> (a, d, Bit 1, Bit 1) -> d

Initilaised block RAM (contents taken from hex file). (Read during write: reads "dont care")

ramDual :: (Bits a, Bits d) => (a, a, d, Bit 1, Bit 1) -> d

Uninitialised simple dual-port block RAM. (Read during write: reads "dont care")

ramDualInit :: (Bits a, Bits d) => String -> (a, a, d, Bit 1, Bit 1) -> d

Initialised simple dual-port block RAM. (Read during write: reads "dont care")

ramTrueDual :: (Bits a, Bits d) => (a, d, Bit 1, Bit 1) -> (a, d, Bit 1, Bit 1) -> (d, d)

Uninitialised true dual-port block RAM. (Read during write: reads "dont care")

ramTrueDualInit :: (Bits a, Bits d) => String -> (a, d, Bit 1, Bit 1) -> (a, d, Bit 1, Bit 1) -> (d, d)

Block RAM primitives with byte enables

ramBE :: KnownNat be => (Bit a, Bit (8 * be), Bit 1, Bit 1, Bit be) -> Bit (8 * be)

Uninitialised block RAM. (Read during write: reads "dont care")

ramInitBE :: KnownNat be => String -> (Bit aw, Bit (8 * be), Bit 1, Bit 1, Bit be) -> Bit (8 * be)

Initilaised block RAM (contents taken from hex file). (Read during write: reads "dont care")

ramDualBE :: KnownNat be => (Bit aw, Bit aw, Bit (8 * be), Bit 1, Bit 1, Bit be) -> Bit (8 * be)

Uninitialised block RAM. (Read during write: reads "dont care")

ramDualInitBE :: KnownNat be => String -> (Bit aw, Bit aw, Bit (8 * be), Bit 1, Bit 1, Bit be) -> Bit (8 * be)

Initialised block RAM (contents taken from hex file). (Read during write: reads "dont care")

ramTrueDualBE :: KnownNat be => (Bit a, Bit (8 * be), Bit 1, Bit 1, Bit be) -> (Bit a, Bit (8 * be), Bit 1, Bit 1, Bit be) -> (Bit (8 * be), Bit (8 * be))

Uninitialised true dual-port block RAM. (Read during write: reads "dont care")

ramTrueDualInitBE :: KnownNat be => String -> (Bit a, Bit (8 * be), Bit 1, Bit 1, Bit be) -> (Bit a, Bit (8 * be), Bit 1, Bit 1, Bit be) -> (Bit (8 * be), Bit (8 * be))

RTL block RAM interface

data RAM a d

RAM interface

Constructors

RAM

Fields

load :: a -> Action ()

Issue load request

store :: a -> d -> Action ()

Issue store request

out :: d

Data out

storeActive :: Bit 1

Is there a store currently happening?

makeRAM :: (Bits a, Bits d) => Module (RAM a d)

Create uninitialised block RAM

makeRAMInit :: (Bits a, Bits d) => String -> Module (RAM a d)

Create block RAM with initial contents from hex file

makeRAMCore :: (Bits a, Bits d) => Maybe String -> Module (RAM a d)

makeTrueDualRAM :: (Bits a, Bits d) => Module (RAM a d, RAM a d)

Create true dual-port block RAM. (Read during write: reads "dont care")

makeTrueDualRAMInit :: (Bits a, Bits d) => String -> Module (RAM a d, RAM a d)

Create true dual-port block RAM with initial contents from hex file. (Read during write: reads "dont care")

makeTrueDualRAMCore :: (Bits a, Bits d) => Maybe String -> Module (RAM a d, RAM a d)

makeDualRAM :: (Bits a, Bits d) => Module (RAM a d)

Create uninitialised dual-port RAM. One port used for reading and the other for writing.

makeDualRAMInit :: (Bits a, Bits d) => String -> Module (RAM a d)

Create dual-port RAM with initial contents from hex file. One port used for reading and the other for writing.

makeDualRAMCore :: (Bits a, Bits d) => Maybe String -> Module (RAM a d)

makeDualRAMForward :: (Bits a, Bits d) => Module (RAM a d)

Uninitialised dual-port forwarding block RAM. Read and write to same address yields new data.

makeDualRAMForwardInit :: (Bits a, Bits d) => String -> Module (RAM a d)

Dual-port forwarding block RAM with initial contents from hex file. Read and write to same address yields new data.

nullRAM :: Bits d => RAM a d

RAM interface with no backing functionality

RTL block RAM interface with byte enables

data RAMBE aw dw

RAM (with byte enables) interface (data width is in bytes)

Constructors

RAMBE

Fields

loadBE :: Bit aw -> Action ()

Issue load request

storeBE :: Bit aw -> Bit dw -> Bit (8 * dw) -> Action ()

Issue store request, with byte enables

outBE :: Bit (8 * dw)

Data out

storeActiveBE :: Bit 1

Is there a store currently happening?

makeRAMBE :: _ => Module (RAMBE aw dw)

Create uninitialised block RAM with byte enables

makeRAMInitBE :: _ => String -> Module (RAMBE aw dw)

Create block RAM with byte enables and initial contents from hex file

makeRAMBECore :: _ => Maybe String -> Module (RAMBE aw dw)

makeTrueDualRAMBE :: _ => Module (RAMBE aw dw, RAMBE aw dw)

Create true dual-port block RAM with byte enables (Read during write: reads "dont care")

makeTrueDualRAMInitBE :: _ => String -> Module (RAMBE aw dw, RAMBE aw dw)

Create true dual-port block RAM with byte enables and initial contents from hex file (Read during write: reads "dont care")

makeTrueDualRAMBECore :: _ => Maybe String -> Module (RAMBE aw dw, RAMBE aw dw)

makeDualRAMBE :: _ => Module (RAMBE aw dw)

Create uninitialised dual-port RAM with byte enables. One port used for reading and the other for writing.

makeDualRAMInitBE :: _ => String -> Module (RAMBE aw dw)

Create dual-port RAM with byte enables and initial contents from hex file. One port used for reading and the other for writing.

makeDualRAMBECore :: _ => Maybe String -> Module (RAMBE aw dw)

makeDualRAMForwardBE :: _ => Module (RAMBE aw dw)

Create uninitialised forwarding dual-port RAM with byte enables. One port used for reading and the other for writing. Read-during-write yields new data.

makeDualRAMForwardInitBE :: _ => String -> Module (RAMBE aw dw)

Create forwarding dual-port RAM with byte enables and initial contents from hex file. One port used for reading and the other for writing. Read-during-write yields new data.

makeDualRAMForwardBECore :: _ => Maybe String -> Module (RAMBE aw dw)

nullRAMBE :: KnownNat (8 * dw) => RAMBE aw dw

RAMBE interface with no backing functionality