Blarney.Core.RTL

Description

The module defines the RTL monad, supporting:

Mutable wires, registers and register files.

Conditional statements.

Simulation-time I/O.

Module input and output declarations.

RTL monad

type RTL = ReaderT RTL_R (StateT RTL_S (WriterT RTL_W Identity))

The RTL monad, for register-transfer-level descriptions

evalPureRTL :: RTL a -> String -> a

evalRTLRoots :: RTL a -> [BV]

Conditional statements

whenRTL :: Bit 1 -> RTL a -> RTL a

RTL conditional block with return value

ifThenElseRTL :: Bits a => Bit 1 -> RTL a -> RTL a -> RTL a

If-then-else statement for RTL

switch :: Bits a => a -> [(a, RTL ())] -> RTL ()

RTL switch statement

(-->) :: a -> RTL () -> (a, RTL ()) infixl 0

Operator for switch statement alternatives

Block naming statements

withNameHint :: NameHint -> RTL a -> RTL a

Set a name hint for an RTL block

Mutable variables: registers and wires

data Reg a

Register variables

Constructors

Reg

Fields

regId :: VarId

Unique identifier

regVal :: a

Current register value

writeReg :: Bits a => Reg a -> a -> RTL ()

Register assignment

data Wire a

Wire variables

Constructors

Wire

Fields

wireId :: VarId

Unique identifier

wireVal :: a

Current wire value

active :: Bit 1

Is wire being assigned on this cycle?

writeWire :: Bits a => Wire a -> a -> RTL ()

Wire assignment

makeReg :: Bits a => a -> RTL (Reg a)

Create register with initial value

makeRegU :: Bits a => RTL (Reg a)

Create wire with don't care initial value

makeDReg :: Bits a => a -> RTL (Reg a)

A DReg holds the assigned value only for one cycle. At all other times, it has the given default value.

makeWire :: Bits a => a -> RTL (Wire a)

Create wire with default value

makeWireU :: Bits a => RTL (Wire a)

Create wire with don't care default value

Simulation-time statements

class Displayable a where

To support a display statement with variable number of arguments

Methods

disp :: Format -> Format -> a

display :: Displayable a => a

Display statement

display_ :: Displayable a => a

Display statement (without new line)

finish :: RTL ()

Terminate simulator

Assertions

assert :: Bit 1 -> String -> RTL ()

Assert that a predicate holds

External inputs and outputs

input :: KnownNat n => String -> RTL (Bit n)

RTL external input declaration

output :: String -> Bit n -> RTL ()

RTL external output declaration

inputBV :: String -> Width -> RTL BV

RTL external input declaration (untyped)

outputBV :: String -> BV -> RTL ()

RTL external output declaration (untyped)

Register files

data RegFileRTL a d

Register file interface

Constructors

RegFileRTL

Fields

lookupRTL :: a -> d

updateRTL :: a -> d -> RTL ()

makeRegFileInit :: forall a d. (Bits a, Bits d) => String -> RTL (RegFileRTL a d)

Create register file with initial contents

makeRegFile :: forall a d. (Bits a, Bits d) => RTL (RegFileRTL a d)

Create uninitialised register file

Add netlist roots

addRoots :: [BV] -> RTL ()

Add netlist roots