Blarney.Misc.ANSIEscapeSequences

Description

This module implements a few helpers to assist with rich text output in blarney. In no way is it aiming to be a complete set of commands. In fact, only some of the Control Sequence Introducer subset of commands are provided, and the limited information used was entirely leveraged from https:/en.wikipedia.orgwiki/ANSI_escape_code.

Control Sequence Introducer

csi_pfx :: String

CSI command prefix

ANSI control sequences

cuu :: Show a => a -> [Char]

cursor up by n

cud :: Show a => a -> [Char]

cursor down by n

cuf :: Show a => a -> [Char]

cursor forward by n

cub :: Show a => a -> [Char]

cursor back by n

cnl :: Show a => a -> [Char]

cursor to start of n-th next line

cpl :: Show a => a -> [Char]

cursor to start of n-th previous line

cha :: Show a => a -> [Char]

cursor to column n

cup :: (Show a, Show a) => a -> a -> [Char]

cursor to row i column j

ed :: Show a => a -> [Char]

erase in display

el :: Show a => a -> [Char]

erase in line

su :: Show a => a -> [Char]

scroll up

sd :: Show a => a -> [Char]

scroll down

hvp :: (Show a, Show a) => a -> a -> [Char]

horizontal vertical pos

sgr :: [[Char]] -> [Char]

select graphic rendition

Select Graphic Rendition options

sgr_rst :: String

reset graphic rendition

sgr_bold :: String

set graphic rendition to bold

sgr_faint :: String

set graphic rendition to faint

sgr_italic :: String

set graphic rendition to italics

sgr_underline :: String

set graphic rendition to underlined

sgr_slow_blink :: String

set graphic rendition to slowly blinking

sgr_rapid_blink :: String

set graphic rendition to rapidly blinking

sgr_reverse :: String

set graphic rendition to swapped foreground and background colors

sgr_conceal :: String

set graphic rendition to hidden

sgr_crossed_out :: String

set graphic rendition to crossed out

sgr_fg_color :: (Ord a, Num a, Show a) => a -> String

set the foreground color

sgr_bg_color :: (Ord a, Num a, Show a) => a -> String

set the background color

sgr_fg_color_8 :: Show a => a -> [Char]

set the 8-bit foreground color, argument value: 0- 7: standard colors (as in ESC [ 30–37 m) 8- 15: high intensity colors (as in ESC [ 90–97 m) 16-231: 6 × 6 × 6 cube (216 colors): 16 + 36 × r + 6 × g + b (0 ≤ r, g, b ≤ 5) 232-255: grayscale from black to white in 24 steps

sgr_bg_color_8 :: Show a => a -> [Char]

same as fg_color_8 for background color

sgr_fg_color_24 :: Show a => a -> a -> a -> [Char]

set the 24-bit foreground color, with 8-bit red + green + blue component

sgr_bg_color_24 :: Show a => a -> a -> a -> [Char]

set the 24-bit background color, with 8-bit red + green + blue component

String wrapper

bold :: [Char] -> [Char]

faint :: [Char] -> [Char]

italic :: [Char] -> [Char]

underline :: [Char] -> [Char]

slow_blink :: [Char] -> [Char]

rapid_blink :: [Char] -> [Char]

crossed_out :: [Char] -> [Char]

red :: [Char] -> [Char]

green :: [Char] -> [Char]

yellow :: [Char] -> [Char]

blue :: [Char] -> [Char]

magenta :: [Char] -> [Char]

cyan :: [Char] -> [Char]

white :: [Char] -> [Char]