Blarney.Misc.MonadLoops

Description

This module provides monadic loop helpers non present in base. Inspired by Control.Monad.Loops

Documentation

untilPredM :: Monad m => m a -> (a -> Bool) -> m [a] infixr 0

Repeat computation until a predicate holds

untilPredM_ :: Monad m => m a -> (a -> Bool) -> m () infixr 0

Same as untilPredM but discard the final result

untilM :: Monad m => m a -> m Bool -> m [a] infixr 0

Repeat computation until the second monadic computation returns True

untilM_ :: Monad m => m a -> m Bool -> m () infixr 0

Same as untilM but discard the final result