Blarney.Netlist

Description

This module exports some utilities to work with Blarney Netlists. It is typically omported by code in modules such as Backend.

Default set of Netlist transformations

runDefaultNetlistPasses :: Opts -> Netlist -> Netlist

Run the default set of netlist passes

Netlist transformation passes (MNetlistPass) helpers

runNetlistPass :: (forall s. MNetlistPass s a) -> Netlist -> Netlist

Run an MNetlistPass on a Netlist and return the resulting Netlist

optionalNetlistPasses :: Opts -> MNetlistPass s ()

wrapWithMandatoryNetlistPasses :: MNetlistPass s a -> MNetlistPass s ()

Wrap a custom pass with the mandatory netlist transformation passes

Exports of individual MNetlistPasses and other utils

module Blarney.Netlist.Passes

others

onNetlists

Arguments

:: Modular a

=> a
Blarney circuit

-> String
circuit name

-> (Map String Netlist -> IO b)
function to run

-> IO b

Run an IO function with the elaborated Netlists hierarchy for the given circuit (effectively one Netlist per eligible Custom Net). The Netlists are passed to the provided function as a 'Map String Netlist', with the toplevel Netlist's key being the provided String argument, and the other Netlists using the $sel:customName:Const field of the elaborated Custom as a key.