Blarney.Netlist.Passes

Description

Documentation

runNetlistPass :: (forall s. MNetlistPass s a) -> Netlist -> Netlist

Run an MNetlistPass on a Netlist and return the resulting Netlist

runDefaultNetlistPasses :: Opts -> Netlist -> Netlist

Run the default set of netlist passes

wrapWithMandatoryNetlistPasses :: MNetlistPass s a -> MNetlistPass s ()

Wrap a custom pass with the mandatory netlist transformation passes

optionalNetlistPasses :: Opts -> MNetlistPass s ()

module Blarney.Netlist.Passes.Utils

module Blarney.Netlist.Passes.ZeroWidthNetIgnore

module Blarney.Netlist.Passes.ConstantFold

module Blarney.Netlist.Passes.ConstantPropagate

module Blarney.Netlist.Passes.ConstantEliminate

module Blarney.Netlist.Passes.NetInline

module Blarney.Netlist.Passes.NamePropagate

module Blarney.Netlist.Passes.DeadNetEliminate