Blarney.Netlist.Passes.Types

Description

This module defines the notions of mutable netlists MNetlists and passes MNetlistPasses which can be used to write Netlist optimisation passes.

Mutable netlist types

type MNetlist s = STArray s InstId (Maybe Net)

A helper type for mutable Netlist

type MNetlistRef s = STRef s (MNetlist s)

A reference to an MNetlist

type MNetlistPass s a = MNetlistRef s -> ST s a

A pass over an MNetlist