Blarney.Netlist.Passes.Utils

Description

This module provides helper functions to write netlist passes.

Queries on individual Nets

netIsRoot :: Net -> Bool

A helper function to tell if a Net is a netlist root

netDontKill :: Net -> Bool

A helper function to tell if a Net should not be optimised away

readNet :: InstId -> MNetlistPass s Net

A helper function to read a Net from a MNetlist

Basic MNetlistPasses

type NetCounts s = STUArray s InstId Int

A helper type for Net reference counting

countNetRef :: MNetlistPass s (NetCounts s)

A Net reference counting helper function

Exported Blarney modules

module Blarney.Core.Prim

module Blarney.Core.NetHelpers