Blarney.Option

Description

An Option type, similar to Maybe in functionality but represented as a pair of a flag and a value.

Option type

data Option t

Option type to wrap a value. An 'Option t' is represented as a pair of a 'Bit 1' indicating whether the value held is valid, and a value of type t.

Constructors

Option

Fields

valid :: Bit 1

val :: t

some :: Bits t => t -> Option t

Builds an Option with a valid value

none :: Bits t => Option t

Builds an invalid Option

isSome :: Bits t => Option t -> Bit 1

Tests if an Option is a some

isNone :: Bits t => Option t -> Bit 1

Tests if an Option is a none

fromOption :: Bits t => t -> Option t -> t

Gets the value from a valid Option, or a given default value otherwise