Blarney.PulseReg

Description

An assignment to a PulseReg is visible only for one cycle, on the cycle after which the assignment is made.

Documentation

data PulseReg

Emits a pulse on the cycle after the pulse method is called

Constructors

PulseReg

Fields

pulse :: Action ()

Trigger pulse on next cycle

val :: Bit 1

Is the register currently pulsing?

makePulseReg :: Module PulseReg