Blarney.Queue

Description

Documentation

data Queue a

Queue interface

Constructors

Queue

Fields

notEmpty :: Bit 1

notFull :: Bit 1

enq :: a -> Action ()

deq :: Action ()

canDeq :: Bit 1

first :: a

clear :: Action ()

makeQueue :: Bits a => Module (Queue a)

A full-throughput 2-element queue implemented using 2 registers:

No combinatorial paths between sides.

There's a mux on the enqueue path.

makeSizedQueue :: Bits a => Int -> Module (Queue a)

A full-throughput N-element queue

data SizedQueueConfig a

Config options for a sized queue

Constructors

SizedQueueConfig

Fields

sizedQueueLogSize :: Int

Log of capacity of queue

sizedQueueBuffer :: Module (Queue a)

Type of queue to use as buffer on output of RAM

makeSizedQueueConfig :: Bits a => SizedQueueConfig a -> Module (Queue a)

Sized queue with config options

makeSizedQueueCore :: Bits a => Int -> Module (Queue a)

This one has no output buffer (low latency, but not great for Fmax)

data ShiftQueueMode

There are modes of operation for the shift queue (below):

Optimise throughput: full throughput, but there's a combinatorial path between notFull and deq

Optimise Fmax: no combinatorial paths between sides, but max throughput = N/(N+1), where N is the queue capacity

Constructors

OptFmax

OptThroughput

makeShiftQueue :: Bits a => Int -> Module (Queue a)

An N-element queue implemented using a shift register:

No muxes: input element goes straight to a register and output element comes straight from a register.

N-cycle latency between enqueuing an element and being able to dequeue it, where N is the queue capacity.

This version optimised for Fmax.

makePipelineQueue :: Bits a => Int -> Module (Queue a)

This version is optimised for throughput

makeShiftQueueCore

Arguments

:: Bits a

=> ShiftQueueMode
Optimise for throughput or Fmax?

-> Int
Size of queue

-> Module (Queue a, [Option a])
Return a standard queue interface, but also the contents to be viewed

makeBypassQueue :: Bits a => Module (Queue a)

Single element bypass queue

makeSinkBuffer :: Bits a => Module (Queue a) -> Sink a -> Module (Sink a)

Insert a queue in front of a sink