Blarney.Recipe

Description

Recipe is a lightweight imperative language, aiding concise definitions of complex state machines.

Documentation

data Recipe

Abstract syntax of Recipe

Constructors

Skip

Tick

Action (Action ())

Seq [Recipe]

Par [Recipe]

Wait (Bit 1)

When (Bit 1) Recipe

If (Bit 1) Recipe Recipe

While (Bit 1) Recipe

Background Recipe

runOnce :: Recipe -> Module ()

Same as runRecipe. For backwards compatibility.

runRecipe :: Recipe -> Module ()

Run a recipe with a start pulse that is high only on the first cycle of execution, and ignore the finish pulse.

runRecipeOn :: Bit 1 -> Recipe -> Module (Bit 1)

Run a recipe, triggered by the given start pulse. Returns the finish pulse.