Blarney.SourceSink

Description

This module defines the Source and Sink types. These types can be used as Module Interfaces do provide a standardized way of handling flow control.

Documentation

data Source t

Source type. A Source t Interface returns values of type t via its $sel:peek:Source method and flow control through its $sel:canPeek:Source and $sel:consume:Source methods.

Constructors

Source

Fields

canPeek :: Bit 1

Checks whether a value can be $sel:peek:Sourceed or $sel:consume:Sourced on this cycle.

peek :: t

Returns the value that the Source is producing this cycle. The value is only relevant on a cycle where $sel:canPeek:Source is true.

consume :: Action ()

Consumes a value from the Source. $sel:consume:Source should only be called on a cycle where $sel:canPeek:Source is true. When implementing a Module returning a Source Interface, one might consider using the signal returned by $sel:canPeek:Source to protect the Action with a when.

data Sink t

Sink type. A Sink t Interface can be fed values of type t via its $sel:put:Sink method, expected to be called when its $sel:canPut:Sink method returns true.

Constructors

Sink

Fields

canPut :: Bit 1

Checks whether a value can be $sel:put:Sink this cycle

put :: t -> Action ()

Puts a value of type t into the Sink

class ToSource a b | a -> b where

ToSource class to convert to a Source

Methods

toSource :: a -> Source b

Converts its argument of type a to a Source b

class ToSink a b | a -> b where

ToSink class to convert to a Sink

Methods

toSink :: a -> Sink b

Converts its argument of type a to a Sink b

nullSource :: Bits t => Source t

"Null" Source that never produces any valid output.

makeNullSource :: Bits t => Module (Source t)

"Null" Source Module that never produces any valid output.

nullSink :: Sink t

"Null" Sink that always consumes its input.

makeNullSink :: Module (Sink t)

"Null" Sink Module that always consumes its input.

mapSource

Arguments

:: (a -> b)
The function to map over the initial Source

-> Source a
The initial Source

-> Source b
The new Source

"map" a function over a Source and returns the new Source. Note: this function is suitable for use as fmap in a Functor instance of Source.

mapSink

Arguments

:: (b -> a)
The function to map over the initial Sink. Note: The type 'b -> a' is making this function not suitable for a potential Functor instance of Sink

-> Sink a
The initial Source

-> Sink b
The new Source

"map" a function over a Sink and returns the new Sink

guardSource :: (a -> Bit 1) -> Source a -> Source a

Apply a guard to a source

debugSource

Arguments

:: FShow t

=> Source t
The Source to debug

-> Format
A Format prefix

-> Source t
The wrapped Source

Wraps a Source t with some debug info.

debugSink

Arguments

:: FShow t

=> Sink t
The Sink to debug

-> Format
A Format prefix

-> Sink t
The wrapped Sink

Wraps a Sink t with some debug info.