Blarney.Stmt

Description

Documentation

data Stmt a

Statement monad, with mondic bind for sequential composition.

tick :: Stmt ()

skip :: Stmt ()

action :: Action () -> Stmt ()

wait :: Bit 1 -> Stmt ()

par :: [Stmt ()] -> Stmt ()

while :: Bit 1 -> Stmt () -> Stmt ()

background :: Stmt () -> Stmt ()

runStmt :: Stmt () -> Module ()

Run a statement with a start pulse that is high only on the first cycle of execution, and ignore the finish pulse.

runStmtOn :: Bit 1 -> Stmt () -> Module (Bit 1)

Run a statement, triggered by the given start pulse. Returns the finish pulse.