Blarney.Stream

Description

Documentation

type Stream a = Source a

Stream interface

data Source t

Source type. A Source t Interface returns values of type t via its peek method and flow control through its canPeek and consume methods.

Constructors

Source

Fields

canPeek :: Bit 1

Checks whether a value can be peeked or consumed on this cycle.

peek :: t

Returns the value that the Source is producing this cycle. The value is only relevant on a cycle where canPeek is true.

consume :: Action ()

Consumes a value from the Source. consume should only be called on a cycle where canPeek is true. When implementing a Module returning a Source Interface, one might consider using the signal returned by canPeek to protect the Action with a when.

toStream :: ToSource a b => a -> Stream b

Convert to a Stream

nullStream :: Bits a => Stream a

Null stream

guardStream :: (a -> Bit 1) -> Stream a -> Stream a

Apply a guard to a stream

type SP t0 t1 = Stream t0 -> Module (Stream t1)

StreamProcessor type

class ToSP a t0 t1 | a -> t0 t1 where

Convert to a StreamProcessor

Methods

toSP :: a -> SP t0 t1