Blarney.TaggedUnion

Documentation

data TaggedUnion (fields :: [Type])

A abstract type for tagged unions. Example usage: > type Either a b = TaggedUnion ["left" ::: a, "right" ::: b]

data TagName (sym :: Symbol)

Wrapper for field name symbols

data (a :: Symbol) ::: (b :: Type)

For specifying the type of a named field of a tagged union

module GHC.OverloadedLabels

isTagged :: forall u name. (IsTaggedUnion u, KnownSymbol name) => u -> TagName name -> Bit 1 infix 7

Does given tagged union have given tag?

is :: forall u name. (IsTaggedUnion u, KnownSymbol name) => u -> TagName name -> Bit 1 infix 7

Shorthand for isTagged

tag :: forall name m u. (IsTaggedUnion u, KnownSymbol name, m ~ GetMemberType u name, Bits m) => TagName name -> m -> u

Construct a value of a tagged union

untag :: (IsTaggedUnion u, KnownSymbol name, m ~ GetMemberType u name, Bits m) => TagName name -> u -> m

Get the value of given field if the tag matches. If tag doesn't match, return dontCare.

untagDefault :: forall name m u. (IsTaggedUnion u, KnownSymbol name, m ~ GetMemberType u name, Bits m) => TagName name -> m -> u -> m

Get the value of given field if the tag matches. If tag doesn't match, return default value.

whenTagged :: (IsTaggedUnion u, KnownSymbol name, m ~ GetMemberType u name, Bits m) => TagName name -> u -> (GetMemberType u name -> Action a) -> Action a

Conditional statement for tagged unions

class IsTaggedUnion t where

Associated Types

type GetMemberType t (tag :: Symbol) :: Type

Methods

getMembers :: t -> [(String, Int)]

toRaw :: t -> (BV, BV)

fromRaw :: (BV, BV) -> t