Blarney.Backend.Simulation

Description

Simulate a Blarney Netlist.

Documentation

simulateNetlist :: Netlist -> IO ()

Simulate a Netlist until a Finish Prim is triggered. This is currently the "only" interface to the simulation module. TODO: export the current SimulatorIfc after sanitizing it, and a bunch of related helpers probably