Blarney.Core.Prim

Description

This module provides a set of circuit primitives.

Prim primitive type and helpers

data Prim

Primitive components

Constructors

Const OutputWidth Integer
Constant value (0 inputs, 1 output)

DontCare OutputWidth
Don't care value (0 input, 1 output)

Add OutputWidth
Adder (2 inputs, 1 output)

Sub OutputWidth
Subtractor (2 inputs, 1 output)

Mul
Multiplier (2 inputs, 1 output). In full-precision mode, output width is 2x input width. Otherwise, input and output width are the same.

Fields

primMulInputWidth :: InputWidth

primMulSigned :: Bool

primMulFullPrecision :: Bool

Div OutputWidth
Quotient (2 inputs, 1 output)

Mod OutputWidth
Remainder (2 inputs, 1 output)

Not OutputWidth
Bitwise not (1 input, 1 output)

And OutputWidth
Bitwise and (2 inputs, 1 output)

Or OutputWidth
Bitwise or (2 inputs, 1 output)

Xor OutputWidth
Bitwise xor (2 inputs, 1 output)

ShiftLeft InputWidth OutputWidth
Left shift (2 inputs, 1 output)

ShiftRight InputWidth OutputWidth
Logical right shift (2 inputs, 1 output)

ArithShiftRight InputWidth OutputWidth
Arithmetic right shift (2 inputs, 1 output)

Equal InputWidth
Equality comparator (2 inputs, 1 bit output)

NotEqual InputWidth
Disequality comparator (2 inputs, 1 bit output)

LessThan InputWidth
Less-than comparator (2 inputs, 1 bit output)

LessThanEq InputWidth
Less-than-or-equal comparator (2 inputs, 1 bit output)

ReplicateBit OutputWidth
Replicate a bit (1 input, 1 output)

ZeroExtend InputWidth OutputWidth
Zero-extend a bit vector (1 input, 1 output)

SignExtend InputWidth OutputWidth
Sign-extend a bit vector (1 input, 1 output)

SelectBits InputWidth BitIndex BitIndex
Bit selection (compile-time range, 1 input, 1 output)

Concat InputWidth InputWidth
Bit vector concatenation (2 inputs, 1 output)

Mux Int OutputWidth
Multiplexer (n inputs, 1 output)

Identity OutputWidth
Identity function (1 input, 1 output)

Register Integer InputWidth
Register with initial value (1 input, 1 output)

RegisterEn Integer InputWidth
Register with initial value and enable wire (2 inputs, 1 output)

BRAM
Block RAM

Fields

ramKind :: BRAMKind

ramInitFile :: Maybe String

ramAddrWidth :: Width

ramDataWidth :: Width

ramHasByteEn :: Bool

Custom
Custom component

Fields

customName :: String

customInputs :: [(String, InputWidth)]

customOutputs :: [(String, OutputWidth)]

customParams :: [Param]

customIsClocked :: Bool

Input OutputWidth String
External input (0 inputs, 1 output)

Output InputWidth String
External output (only used in RTL context, not expression context)

Display [DisplayArg]
Simulation-time I/O (only used in RTL context, not expression context)

Finish
Finish simulation (input: 1-bit enable wire)

TestPlusArgs String
Test presence of plusargs (output: 1-bit boolean)

Assert String
Assertion that a predicate holds (2 inputs, 0 outputs) * String argument is an assertion message * First input is the current RTL condition * Second input is the predicate

RegFileMake RegFileInfo
Register file declaration (only used in RTL context, not expression context)

RegFileRead RegFileInfo
Register file lookup (input: index, output: data)

RegFileWrite RegFileInfo
Register file update (inputs: write-enable, address, data)

canInline :: Prim -> Bool

Helper to tell whether a Prim can be inlined during Netlist optimisation

canInlineInput :: Prim -> Bool

Helper to tell whether a Prim inputs can be inlined during Netlist optimisation

primStr :: Prim -> String

Helper to render a primitive name. Used to generate useful names

primDontKill :: Prim -> Bool

Helper to tell if a Prim can be optimized away or not

primIsRoot :: Prim -> Bool

Helper to tell if a Prim is a netlist root

primInputs :: Prim -> [(String, InputWidth)]

Helper to get the inputs of a Prim

primOutputs :: Prim -> [(String, OutputWidth)]

Helper to get the outputs of a Prim

primOutWidth :: Prim -> OutputName -> OutputWidth

Helper to get the OutputWidth for a given named output of a Prim

data BRAMKind

Kind of block RAM

Constructors

BRAMSinglePort

BRAMDualPort

BRAMTrueDualPort

Other primitive types

type InstId = Int

Every instance of a component in the circuit has a unique id

type Width = Int

Bit vector width

type InputWidth = Width

Width of an input to a primitive

type OutputWidth = Width

Width of an output from a primitive

type OutputName = Maybe String

Reference to a named output of a Prim

type BitIndex = Int

For indexing a bit vector

data RegFileInfo

Register file primitive parameters

Constructors

RegFileInfo

Fields

regFileId :: Int

regFileInitFile :: String

regFileAddrWidth :: InputWidth

regFileDataWidth :: Width

data DisplayArg

For the Display primitive: display a string literal or a bit-vector value of a given width

Constructors

DisplayArgString String
Display a string

DisplayArgBit
Display a bit vector with formatting options

Fields

displayArgWidth :: InputWidth

Radix of bit vector to display

displayArgRadix :: DisplayArgRadix

Radix of bit vector to display

displayArgPad :: Maybe Int

Optional padding

displayArgZeroPad :: Bool

Pad with zeros or spaces?

data DisplayArgRadix

Format for displaying bit vectors

Constructors

Bin

Dec

Hex

data Param

Custom components may have compile-time parameters. A parameter has a name and a value, both represented as strings

Constructors

String :-> String

data NameHint

A NameHint type describing name hints

Constructors

NmPrefix Int String

NmRoot Int String

NmSuffix Int String

type NameHints = Set NameHint

A NameHints type to gather name hints