Blarney.Netlist

Description

This module exports some utilities to work with Blarney Netlists. It is typically omported by code in modules such as Backend.

Default set of Netlist transformations

runDefaultNetlistPasses :: Opts -> Netlist -> Netlist

Run the default set of netlist passes

Netlist transformation passes (MNetlistPass) helpers

runNetlistPass :: (forall s. MNetlistPass s a) -> Netlist -> Netlist

Run an MNetlistPass on a Netlist and return the resulting Netlist

optionalNetlistPasses :: Opts -> MNetlistPass s ()

wrapWithMandatoryNetlistPasses :: MNetlistPass s a -> MNetlistPass s ()

Wrap a custom pass with the mandatory netlist transformation passes

Exports of individual MNetlistPasses and other utils

module Blarney.Netlist.Passes

Orphan instances

ToNetlist (MNetlistRef s) (ST s)
ToNetlist instance for MNetlistRefs in the ST monad

Instance details

Methods

toNetlist :: MNetlistRef s -> ST s Netlist