Blarney.Netlist.Passes.Utils

Description

Documentation

type MNetlist s = STArray s InstId (Maybe Net)

A helper type for mutable Netlist

type MNetlistPass s a = MNetlist s -> ST s a

A type for running a pass over an MNetlist

netIsRoot :: Net -> Bool

A helper function to tell if a Net is a netlist root

netDontKill :: Net -> Bool

A helper function to tell if a Net should not be optimised away

readNet :: InstId -> MNetlistPass s Net

A helper function to read a Net from a MNetlist

type NetCounts s = STUArray s InstId Int

A helper type for Net reference counting

countNetRef :: MNetlistPass s (NetCounts s)

A Net reference counting helper function

untilM :: Monad m => (a -> Bool) -> m a -> m a

Repeat computation until a predicate holds

untilM_ :: Monad m => (a -> Bool) -> m a -> m ()

Same as untilM but discard the final result

module Blarney.Core.Net

module Blarney.Core.Prim